Split resolver replay semantic tests
